minor bug: incorrect handling of split then halt
The cancel_all_orders_for_asset call can induce an assertion failure in the case of a reverse split and delisting. See example below.
In fact, in running the test from January 2015 through November 26, 2015, the following scenario played out.

The stock was ordered on 8/25, 2 shares long
The stock experienced a reverse split on 8/25 with a ratio of 7, making the order amount become 2/7 = 0
The stock was halted on 11/25
The past_auto_close_date test for the asset evaluated to true (Quandl set its end date to 11/24)
The cleanup_expired_assets subroutine was invoked, calling cancel_all on the open order for Noranda
Since the order amount was 0, its status was filled, so it was not deleted
The “assert orders” line failed
If the process finds that the current market data set is missing information about a stock, the order for that stock should be invalidated, but this should be done in a fashion which does not break the execution of assert statements. In fact, though, the check on status of open in the open_orders list is redundant, and thus can be removed without loss of generality.  This would work, if cancel were called only from the cancel_all_orders_asset method in blotter, but, as API calls can be dispatched to this method on any order id, it might be possible to cancel an order twice, updating the timestamp both times. Thus, the correct approach to solve the reverse split issue is likely to cancel the order, if the split causes the amout to become zero.