BUG: Handle rolling to contract with auto-close date outside window
This fixes a bug with the edge case where:

- A contract (C_0) in the chain has an auto-close date equal to the
  `end` date passed to get_rolls
- The contract following C_O (C_1), is *not* the active contract
  relative to the contract after it (C2), e.g. C_1 has lower volume than
  C_2, etc.

In this case, we need to roll from C_2 back to C_0, in which we would
previously jump to the day following C_0's auto-close date. However, in
this edge case, that would be outside the desired window (past `end`).
Instead, in this case, we'll jump to `end` instead.