ENH: --force-reload optional parameter.
zipline run calls function load_market_data. This function has a
cooldown of 1 hour. If it downloaded data less than 1 hour ago it won't
download again.

This has an inconvenient side-effect: if zipline run is executed,
data is not cached, data is downloaded but for some reason data is
still not present after download (like what happens with issue #1965),
if then the issue is fixed, zipline run can still not be executed until
the cooldown time has expired. This can then lead users to open
new issues which seem related to the original one, but are not (see
for example issue #1957 and discussion therein).

This commit adds an optional --force-redownload flag to zipline run.
The default is set to false.