ENH: Added futures bundle
For a user who has access to their own futures data it can be difficult
to figure out how to use that data with zipline. That is because currently
there is no way for a user to ingest futures CSV data into a bundle.

This commit provides that.

This bundle does NOT know how to ingest equities and futures data together.
It takes care of futures data only. In particular it uses the equities binary
writer, but correctly writes futures metadata into the bundle database. The
end result is the desired: a user with futures CSV data will be able to
backtest their futures strategies on zipline.