ENH: Fail better on AssetFinder(nonexistent_path).
Make the AssetFinder fail with a clearer error if you construct and
AssetFinder pointing to a file path that doesnt exist.

The current implementation fails with:

```
InvalidRequestError: Could not reflect: requested table(s) not available in sqlite:///<filepath>
```

which makes it look like the error was that the db was corrupted, when
the actual error was just that the file didn't even exist. This is made
worse by the fact that sqlite3 will actually create an empty file when
you call `connect`, which further propagates the appearance of having
previously had a bad file.