DOC: Add comments explaining each test case.
A test case is an argument that should persuade the reader that the code
under test is correct. It should be easy for a reader to see how the
inputs to a test imply the expected outputs.

I'm often not a fan of `nose.parameterized`, because I think it can make
it easy to write tests where the relationship between the inputs and
outputs isn't clear. One way to mitigate that is to add short comments
to the individual test cases explaining why they should be true.