TEST: Make open different from close.
The expected test outputs are all a fixed percentage away from the close
price, but since open and close were the same value, we wouldn't have
noticed if, for example, we read from the open instead of the
close. Giving each price a distinct value means that the test actually
enforces that we use the correct price.