ENH: Don't dispatch pipeline terms we won't load.
If a loadable term is in the graph but never loaded because all its
dependents are pre-populated, we don't need to dispatch to its loader.

This makes it so that you can avoid setting up all the loaders for a
complex term if you supply it in the initial workspace, which is useful
for local development and testing.