PERF: Use one field when getting the last traded dt.
Most often, when requesting the `last_traded_dt`, it is for retrieving the
forward filled price, which is based on the `close` column.

Instead of using `volume` to get the `last_traded_dt`, use the `close`. This
should particularly help workloads which are retrieving the `price` for many
assets, by reducing the number of I/O operations and making it less likely that
the desired `close` blocks stay in the page cache, since the `volume` and
`close` blocks would not be competing for space.