PERF: Optimize price for liquid assets.
When using `price`, the call to `last_traded_dt` for every value retrieved
became a noticeable bottleneck in algorithms which used over 1000 assets.

Instead of calling `last_traded_dt` before every retrieval of a `close` for the
`price` field, assume that `close` will retrieve a non-empty value, and then
forward fill if it is empty.

This change optimizes for the case of a tradeable universe which is predominately
composed of liquid assets.