BLD: Fetch tags when unshallowing for conda.
The OSX builds are currently failing, and they appear to be failing to fetch
tags, which causes ``git describe`` to fail when invoked internally by
conda. The tentative culprit for this is that the OSX builds are using a newer
git version, which may have stopped fetching tags by default on --unshallow
clones.