ENH: Enable abstract DataSetFamily subclasses
This change enables us to define "abstract" subclasses of DataSetFamily
which implement some set of attributes, but don't represent tangible
families of DataSets, and thus don't have meaningful values for
`extra_dims`.

Previously, all subclasses of DataSetFamily had to define non-empty
`extra_dims`. Subclasses were detected in DataSetFamilyMeta by checking
if the base classes of the class under construction were specified as
anything other than the designated sentinel value: if so, `extra_dims`
was transformed to an OrderedDict and checked for non-emptiness.

With this change, the metaclass instead checks for an `__abstract__`
attribute of the class under construction: if it is found and set to a
truthy value, `extra_dims` (and `_SliceType`) is left unchecked and
unmodified. Regardless of its value, the `__abstract__` attribute is
removed from the constructed class' body: each subclass which wishes to
remain abstract must specify this attribute again.