PERF: Avoid slow exception repr in blaze.
The error raised by blaze when you do a field lookup with a nonexistent key on
an expression shows the repr of the expression, which contains the expression's
entire dshape. For exprs that are part of a very large tree, rendering the
dshape can take a significant amount of time.

Band-aid fix this for now by checking if the field exists before doing the
lookup. This saves 30-40 seconds (!) of startup time on Quantopian research.