MAINT: Rework behaviors shared by ComputableTerm subclasses.
Several behaviors are shared by all ComputableTerm subclasses. The alias() and
downsample() methods, for example, are implemented identically in filter,
factor, and classifier, except for the fact that they return different
types (all of which are subclasses of AliasedMixin/DownsampleMixin).

The isnull() and notnull() methods are also largely duplicated between Factor,
and Classifier.

This PR moves most of these shared behaviors into ComputableTerm. Properties
like _aliased_type and _downsampled_type, instead of being implemented
separately (and almost indentically) in Factor, Filter, and Classifier, are now
implemented in terms of a new method, `_principal_computable_term_type`, which
returns Filter/Factor/Classifier, depending on the runtime type of the
expression.