BUG: window_length=0 term inputs were not being copied.
Some terms mutate their inputs, but by not copying the data before passing them
to a term, the input could be corrupted and then used in other terms that didn't
expect the mutated input.

This change uses the same reference counting mechanism to only copy when needed
that is applied to windowed terms.