FIX: Fix to generalize the before_trading_start call for custom calendars
currently the before_trading_start is hard coded to run at 8:45
local time of 'US/eastern' zone which may not work for other
data bundles of other exchanges in different timezone.
Currently before_trading_start is not invoked for data bundle of BSE
stock exchange with 'XBOM' calander.
This fix generalizes the code to call it 1 hour before the trading open
for each calandars local timezone.
Verified it fixes the issue with XBOM calandar bundles.